% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medium01.R
\name{addQualityFlag}
\alias{addQualityFlag}
\title{Function for Retrieving 5000 georeferenced records of Australian mammals from GBIF and adding quality flags.}
\usage{
addQualityFlag(country = "AU", hasCoordinate = TRUE, limit = 5000,
  classKey = 121)
}
\arguments{
\item{country}{The 2-letter country code (as per ISO-3166-1)}

\item{hasCoordinate}{A number}

\item{limit}{Number of records to return}

\item{classKey}{Class classification key}
}
\value{
A tidy dataframe of occurancedata and quality flags of fields as specified in parameters.
}
\description{
{
GSOC
Integrating Biodiversity Data Curation Functionality.
Test Medium 01.
}
}
\examples{
addQualityFlag()
addQualityFlag("SL", False, 100, 148)
}
\author{
Thiloshon Nagarajah
}
