% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsf_medium.R
\name{jsf_simple_simulate}
\alias{jsf_simple_simulate}
\title{Run a simple JSF simulation}
\usage{
jsf_simple_simulate(x0 = c(10, 0), t_max = 10)
}
\arguments{
\item{x0}{Initial state vector, default c(10, 0)}

\item{t_max}{Maximum simulation time, default 10.0}
}
\value{
A list containing simulation results
}
\description{
Core function: calls Python jsf package via reticulate
This implements the basic requirement of Medium Test
}
