% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsfs7.R
\name{JSFResult}
\alias{JSFResult}
\title{JSFResult S7 class}
\usage{
JSFResult(time = integer(0), populations = list(), metadata = list())
}
\description{
Bonus feature: S7 object system encapsulation
Demonstrates understanding of the complete project architecture
}
