% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medium.R
\name{occAM}
\alias{occAM}
\title{occurence data from GBIF datasets.}
\usage{
occAM(country = "AU", classKey = 359, minimal = FALSE)
}
\arguments{
\item{country}{(character) ISO 3166-1-alpha-2 code of a particular country to
get occurences of that country.Default:"AU"}

\item{classKey}{(integer) Key of a particular class(eg. animalia) according to
GBIF. Default:359}

\item{minimal}{(logical) False to get full set of fields for each record.
Default:FALSE}
}
\value{
a data frame with occurence data
}
\description{
function to retrieve occurence data of terrestial
animals in Australia and apply GQ API
}
\details{
function to retrieve occurence data of terrestial animals in Australia and apply GQ API
}
\author{
Vandit Jain
}

